class 별 단위 테스트 파일 분리 하면서 테스트 목록 수정
